# Check for most common case where there is no second header