This is a very minor speed improvement, but it is measurable. Especially on reads with only 43bp, where records are small and the +lines make up relatively more of the file. The difference is still measurable on 150bp given a reasonably large file. On 9000 bp it does not matter.

before:
```
Benchmark 1: python dnaio_read.py ../dnaio-benchmarks/data/single43bp.fastq
  Time (mean ± σ):      1.240 s ±  0.034 s    [User: 1.154 s, System: 0.086 s]
  Range (min … max):    1.199 s …  1.307 s    10 runs
```

After
```
Benchmark 1: python dnaio_read.py ../dnaio-benchmarks/data/single43bp.fastq
  Time (mean ± σ):      1.179 s ±  0.035 s    [User: 1.084 s, System: 0.095 s]
  Range (min … max):    1.128 s …  1.258 s    10 runs
```

I also tested it on my `allinc` branch. The difference is smaller there, but since the C branch has smaller run-to-run variance it is still clearly visible. 

Can you reproduce this on your system? If so, it might be nice to put it in before 0.10.0 to have it count for the benchmarks in the paper. 

For other performance improvements I checked whether I could improve cache locality by rearranging the pointer dereferences. The code is now ordered nicely by function, which makes it more readable, but not by memory access. So ideally you want to check for `@` first, then search the newline, then check for `\r` before the newline etc. That way you don't perform jumps. So I did a small test, but it does not matter. Even not on the 9000bp file. Maybe I improved cache locality, but we keep a max of 1 record in cache. And L1 cache latency is 3-4 clockticks. So that is why it does not matter probably. So it is better to keep it this way, where the code is much more readable.

I also want to see if we can use more knowledge of the format to parse even faster. I think the `+\n` case is a good example. Another case is the qualities line. The qualities line should be the same length as the sequence line. Still we scan for `\n`. But we have to do that, because when files are concatenated (quite common) it might occur that a truncated file concatenated to another file yields a valid result if we do not do the check properly. Still we do know where the `\n` should be and I am hoping that we find a way to utilize that knowledge to make it faster. It will be hard, because `memchr` is basically hand-written assembler and it uses sse2, avx2 or avx512 dependening on a runtime check on your system. It checks one cacheline (64B or 512bits) at the same time. So it is incredibly hard to beat.