@rhpvorderman 2022-11-16
I did a check if I could do something similar to #14 on a `uselength` branch to improve quality line speed, but I have to conclude that even if I did something safely at 0 cost (which is impossible) the gains would be extremely small. On the same magnitude as this PR. Which is not great. Even when reads have 9000 phred scores each, which are all scanned for '\n'. The scanning is so fast, that the actual overhead of dereferencing the `memchr` function pointer accounts for most of the cost. (As shown in this PR, there is a cost to calling memchr as opposed to not doing it). So I think by this point it is almost impossible to improve the parsing code further and most of the gains can be produced by reducing the C to Python overhead.