@marcelm 2022-12-02
I’ve tried to measure this, but don’t have a more or less unused system available for this at the moment, so the variance is too high and I couldn’t see a difference. I’ll just merge this.

Looking at your efforts, I wonder if there’s some great idea out there that would improve this even further significantly ...

Ok to release 0.10.0?