@rhpvorderman 2022-12-05
> so the variance is too high and I couldn’t see a difference.

It really only helps quite short reads. But at least it does not make it slower in any of the other cases.

> Ok to release 0.10.0?

Yes! I am looking forward to it! Especially with all the updates to the open function.

Another note: I have been quite busy lately, but I am planning to work on the dnaio graphs for the paper during the holiday.