@marcelm 2022-12-06
Ok, I’ve made the release. Nice round number :smile:.

> I have been quite busy lately, but I am planning to work on the dnaio graphs for the paper during the holiday.

Same here, I hope I can set aside a couple of hours.